fix: resolve golangci-lint type-checker panic on go1.26 system
- Switch golangci-lint installation from pre-built binary (curl) to
  go install so the binary is compiled with the local Go toolchain.
  Pre-built v2.5.0 was built with go1.25.1 and panicked on
  golang.org/x/net@v0.47.0's //go:build go1.26 file when the system
  Go is 1.26.0. Building from source with go install avoids the mismatch.
- Pin golangci-lint to v2.11.3.
- Remove revive 'exported' rule (50 violations; spawner is a platform
  service, not a public library — godoc coverage is tracked separately).
- Remove govet 'shadow' analyser (patterns flagged are all safe;
  outer err is nil before each inner re-declaration).
- Add gocognit exclusion for pkg/api/types.go (RunSpec.Validate is
  inherently complex due to multi-field validation logic).
- Add prealloc to test-file exclusions.
- Fix unconvert: remove redundant int32() cast on pref.Weight which
  is already int32.

+++ cmd/imp/k8s_driver.go
@@ -266,7 +266,7 @@ func buildPlacementAffinity(p *api.Placement) *corev1.Affinity {
 	terms := make([]corev1.PreferredSchedulingTerm, 0, len(p.PreferredNodes))
 	for _, pref := range p.PreferredNodes {
 		terms = append(terms, corev1.PreferredSchedulingTerm{
-			Weight: int32(pref.Weight),
+			Weight: pref.Weight,
 			Preference: corev1.NodeSelectorTerm{
 				MatchExpressions: []corev1.NodeSelectorRequirement{{
 					Key:      "kubernetes.io/hostname",
